% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{EHDENUseCase6.3AgeSex-package}
\alias{EHDENUseCase6.3AgeSex}
\alias{EHDENUseCase6.3AgeSex-package}
\title{EHDENUseCase6.3AgeSex: A Package Skeleton for Comparative Effectiveness Studies}
\description{
A skeleton package, to be used as a starting point when implementing comparative effect studies.
}
\keyword{internal}
