% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{EHDENUseCase7.2RemTocNoSubgroupsCCISex-package}
\alias{EHDENUseCase7.2RemTocNoSubgroupsCCISex}
\alias{EHDENUseCase7.2RemTocNoSubgroupsCCISex-package}
\title{EHDENUseCase7.2RemTocNoSubgroupsCCISex: A Package Skeleton for Comparative Effectiveness Studies}
\description{
A skeleton package, to be used as a starting point when implementing comparative effect studies.
}
\keyword{internal}
