% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{EHDENUseCase4.3-package}
\alias{EHDENUseCase4.3}
\alias{EHDENUseCase4.3-package}
\title{EHDENUseCase4.3: A Package Skeleton for Comparative Effectiveness Studies}
\description{
A skeleton package, to be used as a starting point when implementing comparative effect studies.
}
\keyword{internal}
