% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{EHDENUseCase7.1RemTocNoSubgroups-package}
\alias{EHDENUseCase7.1RemTocNoSubgroups}
\alias{EHDENUseCase7.1RemTocNoSubgroups-package}
\title{EHDENUseCase7.1RemTocNoSubgroups: A Package Skeleton for Comparative Effectiveness Studies}
\description{
A skeleton package, to be used as a starting point when implementing comparative effect studies.
}
\keyword{internal}
