% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{EHDENUseCase6.4CCISex-package}
\alias{EHDENUseCase6.4CCISex}
\alias{EHDENUseCase6.4CCISex-package}
\title{EHDENUseCase6.4CCISex: A Package Skeleton for Comparative Effectiveness Studies}
\description{
A skeleton package, to be used as a starting point when implementing comparative effect studies.
}
\keyword{internal}
