% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{EHDENUseCase5.2AgeSex-package}
\alias{EHDENUseCase5.2AgeSex}
\alias{EHDENUseCase5.2AgeSex-package}
\title{EHDENUseCase5.2AgeSex: A Package Skeleton for Comparative Effectiveness Studies}
\description{
A skeleton package, to be used as a starting point when implementing comparative effect studies.
}
\keyword{internal}
