% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{EHDENUseCase5.3CCISex-package}
\alias{EHDENUseCase5.3CCISex}
\alias{EHDENUseCase5.3CCISex-package}
\title{EHDENUseCase5.3CCISex: A Package Skeleton for Comparative Effectiveness Studies}
\description{
A skeleton package, to be used as a starting point when implementing comparative effect studies.
}
\keyword{internal}
