% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{EHDENUseCase3.2-package}
\alias{EHDENUseCase3.2}
\alias{EHDENUseCase3.2-package}
\title{EHDENUseCase3.2: A Package Skeleton for Comparative Effectiveness Studies}
\description{
A skeleton package, to be used as a starting point when implementing comparative effect studies.
}
\keyword{internal}
